Added an HTTPie transport plugin
The new contrib plugin enables the direct use of pykube as the transport
layer for HTTPie.

For example:

	pip install httpie pykube
	http pykube://minikube

The above example will enable HTTPie send an HTTP request to the kubeconfig
context named minikube (dealing with certificates and authentication for you).

This isn't 100% compatible with pykube credential handling yet. Further refactoring
is needed to support Google OAuth.